Implement UI config flow
This implementation does not allow for editing the config entry.
It also does not implement migration of the YAML config entry to
the new UI config entry, but since data is stored by `scups`, removing
it from configuration.yaml and setting it up from the UI should work.

`debug` config is still taken from yaml config, since it is global and not per
CUPS.